qol: restyle user level badges and treat user levels as an eligibility floor
